+++ lib/AnimationStation/src/AnimationStation.hpp
@@ -32,14 +32,14 @@ public:
   void Clear();
   void ChangeAnimation();
 
-  static void SetBrightness(int brightness);  
+  static void SetBrightness(uint8_t brightness);  
   static void DecreaseBrightness();
   static void IncreaseBrightness();
   static uint32_t RGB(uint8_t r, uint8_t g, uint8_t b);
   static uint32_t Wheel(uint8_t pos);
 
   std::vector<Animation*> animations;
-  static int brightness;
+  static uint8_t brightness;
   static absolute_time_t nextBrightnessChange;
   static absolute_time_t nextAnimationChange;
   uint32_t frame[100];
